bda-tkwy: drop unnecessary uintptr conversion (unconvert, GOOS=windows lane)
^uintptr(0) is already a uintptr; the wrapper-timing PR Lint lane lints
cross-platform and was the only job still red on a deterministic finding.
GOOS=windows build + golangci-lint run both clean after.

+++ cmd/bd/console_windows.go
@@ -29,7 +29,7 @@ import "syscall"
 var (
 	kernel32          = syscall.NewLazyDLL("kernel32.dll")
 	procAttachConsole = kernel32.NewProc("AttachConsole")
-	attachParentPID   = uintptr(^uintptr(0)) // (DWORD)-1
+	attachParentPID   = ^uintptr(0) // (DWORD)-1
 )
 
 func attachParentConsole() {
